+++ packages/evm-module/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @avalabs/evm-module
 
+## 0.0.15
+
+### Patch Changes
+
+- f536d58: add module functions to evm-module
+- Updated dependencies [f536d58]
+  - @avalabs/vm-module-types@0.0.15
+  - @internal/utils@0.0.3
+
 ## 0.0.14
 
 ### Patch Changes